Fix mask test script

+++ scripts/maskTest.py
@@ -22,17 +22,17 @@ mask = np.zeros((h, w, 3), dtype="uint8")
 mask = cv2.bitwise_not(mask)
 
 #mask[cv2.Range(0, mask.rows), cv2.Range(0, 9 * mask.cols / 240 )] = 0
-cv2.rectangle(mask, (0, 0), (w * 9 / 240, h), (0, 0, 0), -1)
+cv2.rectangle(mask, (0, 0), (int(w * 9 / 240), h), (0, 0, 0), -1)
 #mask[cv2.Range(0, mask.rows / 3), cv2.Range( 9 * mask.cols / 240 , 33 * mask.cols / 240)] = 0
 #mask[cv2.Range(41 * mask.rows / 60, mask.rows), cv2.Range( 9 * mask.cols / 240,  33 * mask.cols / 240)] = 0
 #mask[cv2.Range(0, mask.rows), cv2.Range(33 * mask.cols / 240, mask.cols / 6)] = 0
-cv2.rectangle(mask, (w * 33 / 240, 0), (w * 40 / 240, h), (0, 0, 0), -1)
+cv2.rectangle(mask, (int(w * 33 / 240), 0), (int(w * 40 / 240), h), (0, 0, 0), -1)
 #mask[cv2.Range(0, mask.rows), cv2.Range(mask.cols / 4, 9 * mask.cols / 24)] = 0
-cv2.rectangle(mask, (w * 60 / 240, 0), (w * 90 / 240, h), (0, 0, 0), -1)
+cv2.rectangle(mask, (int(w * 60 / 240), 0), (int(w * 90 / 240), h), (0, 0, 0), -1)
 #mask[cv2.Range(0, mask.rows), cv2.Range(29 * mask.cols / 60, 73 * mask.cols / 120)] = 0
-cv2.rectangle(mask, (w * 116 / 240, 0), (w * 146 / 240, h), (0, 0, 0), -1)
+cv2.rectangle(mask, (int(w * 116 / 240), 0), (int(w * 146 / 240), h), (0, 0, 0), -1)
 #mask[cv2.Range(0, mask.rows), cv2.Range(3 * mask.cols / 4, mask.cols)] = 0
-cv2.rectangle(mask, (w * 180 / 240, 0), (w, h), (0, 0, 0), -1)
+cv2.rectangle(mask, (int(w * 180 / 240), 0), (w, h), (0, 0, 0), -1)
 
 
 #mask1 = cv2.inRange(self.hsv, self.dim["foosballMin1HSV"], self.dim["foosballMax1HSV"])
